reorganize plot functions
